% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C4EstimateWithoutCA.R
\name{C4EstimateWithoutCA}
\alias{C4EstimateWithoutCA}
\title{Estimate photosynthesis parameters for C4 species using Sharkey's fitting procedure}
\usage{
C4EstimateWithoutCA(ACi, Tleaf, Patm, alpha1, x, CaBreakL, CaBreakH,
  startp)
}
\arguments{
\item{ACi}{Gas exchange measurement from Li6400 or other equipment. It is a
dataframe iput. Ci with the unit of ppm. You can prepare the data in Excel file
like the given example and save it as "tab delimited text". Then import data by
ACi <- read.table(file = "/Users/haoranzhou/Desktop/Learn R/ACi curve.txt",header
= TRUE)}

\item{Tleaf}{Leaf temperature when A_Ci curve is measured.}

\item{Patm}{Atmosphere pressure when A_Ci curve is measured.}

\item{alpha1}{The fraction of O2 evolution occurring in the bundle sheath. Unless
you have enough information, input it as the 0.15.}

\item{x}{the fraction of total electron transport that are confined to be
used for the PEP regeneration out of J, which is the total electron transport.}

\item{CaBreakL}{Higher bound of Ci below which A is thought to be controled by
Rubisco Carboxylation (Ac). Start with 10.}

\item{CaBreakH}{Lower bound of Ci above which A is thought to be controled by RuBP
regeneration (Aj). Start with 50. If the estimation results showed "inadmissible
fits", change the CaBreakL and CaBreakH until "inadmissible fits" disappear.}

\item{startp}{A vector that gives the start points for the estimation (c(Vcmax,
J,Rd,gm and Vpmax))}
}
\value{
This package will return a dataframe that contains the following values
(c(Vcmax,J,Rd,gm and Vpmax)). You can try with c(30, 150, 3, 10, 50).

Parameter at leaf temperature:      A vector (c(Vcmax,J,Rd,gm and Vpmax))
returns the estimation parameters at leaf temperature.

Parameter at 25°C:                  A vector (c(Vcmax,J,Rd,gm and Vpmax))
returns the estimation parameters at leaf temperature.

Objective:                          The final objective value based on the
estimation results.

Convergence:                        An integer code. 0 indicates successful
convergence.

Message:	                          A character string giving any additional
information returned by the optimizer, or NULL. For details, see PORT documentation.

Iterations:	                        Number of iterations performed.

Evaluations:	                      Number of objective function and gradient
function evaluations.
}
\description{
Using the gas exchange measurement (A_Ci curve), C4 photosynthesis model without
carbonic anhydrase and Sharkey et al. (2007) fitting processure to do nonlinear
curve fitting (using nlminb package) for estimating photosynthesis parameters
(Vcmax,J,Rd,gm and Vpmax) for C4 species. Make sure to load the "stats" package
before intstalling and using the "C4Estimation" package.
}
